georgebot/bot/requirements.txt
requests
python-dotenv
python-telegram-bot
schedule
scikit-learn
pandas
fastapi
uvicorn
httpx
solana
aiohttp
